Fix concurrency issue with unit tests
Because we cannot change the Items property or type and it is exposed as a static property in the Program class we need to prevent its state being shared between the tests.

Ideally, the concept of current inventory would be injected into the Program class and its value mocked out when running under a test context. Once this is done we no longer need a static property and can undo this change.